+++ src/components/ExperimentalFilters/EventEmitter.ts
@@ -1,112 +1,20 @@
-import { Option } from "..";
-
-export type RightOperatorOption = Option & {
-  slug: string;
-};
-
-export type LeftOperatorOption = Option & {
-  type: string;
-};
-
-export type ConditionOption<T extends string> = Option & {
-  type: T;
-};
-
-export type RowAddData = {
-  type: "row.add";
-  rowType: string;
-};
-
-export type RowRemoveData = {
-  type: "row.remove";
-  path: `${number}`;
-  index: number;
-};
-
-export type LeftOperatorChangeData = {
-  type: "leftOperator.onChange";
-  path: `${number}`;
-  index: number;
-  value: LeftOperatorOption;
-  rowType: string;
-};
-
-export type LeftOperatorFocusData = {
-  type: "leftOperator.onFocus";
-  path: `${number}`;
-  index: number;
-};
-
-export type LeftOperatorBlurData = {
-  type: "leftOperator.onBlur";
-  path: `${number}`;
-  index: number;
-};
-
-export type LeftOperatorInputValueChangeData = {
-  type: "leftOperator.onInputValueChange";
-  path: `${number}`;
-  index: number;
-  value: string;
-};
-
-export type ConditionChangeData = {
-  type: "condition.onChange";
-  path: `${number}.condition.selected`;
-  index: number;
-  value: ConditionOption<
-    | "text"
-    | "number"
-    | "multiselect"
-    | "combobox"
-    | "select"
-    | "number.range"
-    | "date"
-    | "datetime"
-  >;
-};
-
-export type ConditionFocusData = {
-  type: "condition.onFocus";
-  path: `${number}.condition.selected`;
-  index: number;
-};
-
-export type ConditionBlurData = {
-  type: "condition.onBlur";
-  path: `${number}.condition.selected`;
-  index: number;
-};
-
-export type RightOperatorChangeData = {
-  type: "rightOperator.onChange";
-  path: `${number}.condition.selected.value`;
-  index: number;
-  value:
-    | string
-    | RightOperatorOption[]
-    | RightOperatorOption
-    | [string, string];
-};
-
-export type RightOperatorFocusData = {
-  type: "rightOperator.onFocus";
-  path: `${number}.condition.selected.value`;
-  index: number;
-};
-
-export type RightOperatorBlurData = {
-  type: "rightOperator.onBlur";
-  path: `${number}.condition.selected.value`;
-  index: number;
-};
-
-export type RightOperatorInputValueChangeData = {
-  type: "rightOperator.onInputValueChange";
-  path: `${number}.condition.selected.value`;
-  index: number;
-  value: string;
-};
+import {
+  ConditionBlurData,
+  ConditionChangeData,
+  ConditionFocusData,
+  LeftOperatorBlurData,
+  LeftOperatorChangeData,
+  LeftOperatorFocusData,
+  LeftOperatorInputValueChangeData,
+  LeftOperatorOption,
+  RightOperatorBlurData,
+  RightOperatorChangeData,
+  RightOperatorFocusData,
+  RightOperatorInputValueChangeData,
+  RightOperatorOption,
+  RowAddData,
+  RowRemoveData,
+} from "./types";
 
 export class FilterEventEmitter extends EventTarget {
   type = "filterChange";
