+++ src/components/ExperimentalFilters/index.ts
@@ -1,12 +1,5 @@
-export type { FilterEvent } from "./useEvents";
 export type { ExperimentalFiltersProps } from "./Root";
-export type { SelectedOperator } from "./RightOperator";
-export type { RowComponent } from "./Row";
-export type {
-  RightOperatorOption,
-  LeftOperatorOption,
-  ConditionOption,
-} from "./EventEmitter";
+export type * from "./types";
 
 import { Root } from "./Root";
 import { AddRowButton, ConfirmButton, Footer, ClearButton } from "./Footer";
